WEBAPP-234: Move i18nRedirect
