+++ src/routes/i18nRedirect/containers/I18nRedirect.js
@@ -4,11 +4,11 @@ import React from 'react'
 import PropTypes from 'prop-types'
 import { redirect } from 'redux-first-router'
 import { connect } from 'react-redux'
-import CityModel from '../../endpoint/models/CityModel'
-import { goToLanding } from '../routes/landing'
-import { goToCategories } from '../routes/categories'
-import { MAX_LANGUAGE_CODE_LENGTH, MIN_LANGUAGE_CODE_LENGTH } from '../constants'
-import { goToNotFound } from '../routes/notFound'
+import CityModel from '../../../modules/endpoint/models/CityModel'
+import { goToLanding } from '../../../modules/app/routes/landing'
+import { goToCategories } from '../../../modules/app/routes/categories'
+import { MAX_LANGUAGE_CODE_LENGTH, MIN_LANGUAGE_CODE_LENGTH } from '../../../modules/app/constants'
+import { goToNotFound } from '../../../modules/app/routes/notFound'
 
 import type { Action } from 'redux-first-router/dist/flow-types'
